Refactored component, Added  prop - Close #17
