+++ .editorconfig
@@ -1,0 +1,11 @@
+# EditorConfig - http://EditorConfig.org
+
+root = true
+
+[*]
+end_of_line = lf
+insert_final_newline = true
+
+[**.js]
+indent_style = space
+indent_size = 4
